Rename normal camera to standard camera; closes #88 (#91)
