+++ src/styles/camera/standardCamera.js
